MAGECLOUD-919: Refactor ECE Tools
